create new add task container and ui components
