+++ src/components/AddTaskUI.tsx
@@ -1,0 +1,36 @@
+import { ChangeEvent } from 'react';
+
+interface AddTaskUIProps {
+  isBlank: boolean;
+  handleTaskInput: (e: ChangeEvent<HTMLInputElement>) => void;
+  newTask: string;
+}
+
+function AddTaskUI({ isBlank, handleTaskInput, newTask }: AddTaskUIProps) {
+  return (
+    <>
+      <div className="relative w-full">
+        <input
+          className={`w-full border rounded-md p-2 
+          ${isBlank ? 'border-red-600' : 'border-black'}`}
+          placeholder="New task.."
+          type="text"
+          name="newTask"
+          id="newTask"
+          onChange={handleTaskInput}
+          value={newTask}
+        />
+        {isBlank && (
+          <div className="absolute text-sm md:text-base top-full text-red-600">
+            Please fill in a description for the new task.
+          </div>
+        )}
+      </div>
+      <button className="bg-cyan-300 hover:bg-cyan-100 rounded-md border border-black md:w-24 p-1">
+        Add
+      </button>
+    </>
+  );
+}
+
+export default AddTaskUI;
